requirements.txt
pytest==7.1.3
web3==5.31.1
requests==2.28.1
requests_toolbelt==0.10.0
python-mcs-sdk==0.1.1
python-dotenv
pyroute2
